refactor: split the code into several files
