+++ config.py
@@ -1,0 +1,5 @@
+# Замените "YOUR_BOT_TOKEN" на токен, который вы получили от BotFather
+API_TOKEN = 'YOUR_BOT_TOKEN'
+
+# Зададим имя базы данных
+DB_NAME = 'quiz_bot.db'
