+++ quiz_data.py
@@ -1,0 +1,14 @@
+# Структура квиза
+quiz_data = [
+    {
+        'question': 'Что такое Python?',
+        'options': ['Язык программирования', 'Тип данных', 'Музыкальный инструмент', 'Змея на английском'],
+        'correct_option': 0
+    },
+    {
+        'question': 'Какой тип данных используется для хранения целых чисел?',
+        'options': ['int', 'float', 'str', 'natural'],
+        'correct_option': 0
+    },
+    # Добавьте другие вопросы
+]
